# swagger.yaml
openapi: 3.0.0
info:

  title: Jobs Api

  contact: {}

  version: '1.0'
servers:

- url: https://jobs-api-heroku-06.herokuapp.com/

  variables: {}
paths:
...
  /jobs:

    post:

      tags:

      - Jobs

      summary: Create Job

      operationId: CreateJob

      parameters: []

      requestBody:

        description: ''

        content:

          application/json:

            schema:

              $ref: '#/components/schemas/CreateJobRequest'

            example:

              company: peter 2

              position: peter 2

        required: true

      responses:

        '200':

          description: ''

          headers: {}

      deprecated: false

      security:

      - bearer: []

    get:

      tags:

      - Jobs

      summary: Get All Jobs

      operationId: GetAllJobs

      parameters: []

      responses:

        '200':

          description: ''

          headers: {}

      deprecated: false

      security:

      - bearer: []
...
components:
  schemas:
...
    CreateJobRequest:

      title: CreateJobRequest

      required:

      - company

      - position

      type: object

      properties:

        company:

          type: string

        position:

          type: string

      example:

        company: peter 2

        position: peter 2
...
  securitySchemes:

    bearer:

      type: http

      scheme: bearer
security: []
...